// src/models/Customer.tsx
export class Customer {
    name?: string
    socialReason?: string
    fantasyName?: string
    cpF_CNPJ?: string
    owner?: string
    responsible?: string
    address?: string
    city?: string
    zipcode?: string
    state?: string
    phone?: string
    email?: string
    url_image?: string
    customerTypeId?: number
    veterinaryId?: number
    password?: string
    id?: number
}