// src/models/Pet.tsx
export interface Pet {
    name: string
    specie: string
    breed: string
    color: string
    gender: string
    weight: string
    port: string
    birthday: string
    coat: string
    castration: string
    temper: string
    url_image: string
    petsTypeId: number
    customerId: number
    id: number
}